requirements.txt
flask==3.0.0
pyaudio==0.2.13
wave==0.0.2
google-cloud-speech==2.21.0
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0 